elastic web application/requirements.txt
Flask~=2.2.2
plotly~=5.10.0
pandas~=1.3.4
requests~=2.26.0
mysql-connector-python~=8.0.30
APScheduler~=3.9.1